Add crude display window drawing to VIC
This is a really basic implementation that doesn't give a damn about timing and
similar stuff, all it cares is to display pixels where they should appear.
Also, the Memory interface got split into Read and Write, since we VIC doesn't
really need to write anything and this started to become a source of
unnecessary boilerplate code.